refactor: move mouse events to core/common folder
