+++ src/core/common/events/mouse.tsx
@@ -1,0 +1,11 @@
+
+export interface MouseClickEvent {
+    targetX: number;
+    targetY: number;
+    button: number;
+}
+
+export interface MouseDragEvent extends MouseClickEvent {
+    deltaX: number;
+    deltaY: number;
+}
