+++ src/core/common/index.tsx
@@ -2,3 +2,4 @@ export * from './FileSystemRepository';
 export * from './uuidToRgb';
 export * from './dependencyInjection';
 export * from './utils';
+export * from './events';
