+++ src/layouts/EditorLayout/index.tsx
@@ -9,7 +9,8 @@ import { FileSystemSceneRepository } from '../../core/scene/adapters';
 import { FileSystemProjectRepository } from '../../core/project/adapters';
 import { SaveProjectUseCase } from '../../core/project/usecases';
 import { Project } from '../../core/project/models';
-import { MouseClickEvent, MouseDragEvent, OnEngineViewReadyCBProps } from '../../components/EngineView';
+import { OnEngineViewReadyCBProps } from '../../components/EngineView';
+import { MouseClickEvent, MouseDragEvent } from '../../core/common/events/mouse';
 import { FileSystemImageRepository } from '../../core/assets/image/adapters';
 import { WeakRef } from '../../core/common';
 import { ImageRepository } from '../../core/assets';
